Seperated files and made buttons class based
Along with many other changes such as making buttons change colors, more buttons, and sum thingies.
Idea: have the background be a bunch of low opacity, completely grey buttons that fill in, and then have prestige buttons out in the grid that you can get to. 
Another idea: make it so that when buttons change colour, they change to the colour of a pixel of an image, when you get to the end and zoom out it makes the full image.